+++ platformio.ini
@@ -29,6 +29,27 @@ lib_deps =
 ;; when pushing firmware, short GPIO0 and GND pin
 ;; if firmware upload fails, then push the boot button once
 
+[env:esp32CarWifi]
+platform = espressif32
+board = esp32cam
+board_build.partitions = partitions/4mb_spiffs.csv
+board_build.flash_size = 4MB
+board_build.flash_mode = qio
+; change microcontroller
+board_build.mcu = esp32
+framework = arduino
+monitor_speed = 115200
+
+build_src_filter = +<*.h> +<esp32CarWifiWithCamera/*.h> +<esp32CarWifiWithCamera/main.cpp>
+lib_deps = 
+	ESP32Async/AsyncTCP
+	ESP32Async/ESPAsyncWebServer
+	cytrontechnologies/Cytron Motor Drivers Library@^1.0.1
+	madhephaestus/ESP32Servo@^3.0.9
+	waspinator/AccelStepper@^1.64
+
+
+
 [env:esp32CarBleWithSteering]
 platform = espressif32
 board = upesy_wroom
